requirements.txt
#Core dependencies
numpy>=1.21.0,<2.0.0
networkx>=2.6.0,<3.0.0
scipy>=1.7.0,<2.0.0
pandas>=1.3.0,<2.0.0

#Visualization
matplotlib>=3.4.0,<4.0.0
seaborn>=0.11.0,<0.13.0

#Configuration and logging
pyyaml>=5.4.0,<7.0.0
python-json-logger>=2.0.0,<3.0.0

#Testing
pytest>=6.2.0,<8.0.0
pytest-cov>=2.12.0,<5.0.0
pytest-mock>=3.6.0,<4.0.0  # Added for mocking in tests
pytest-timeout>=2.0.0,<3.0.0  # Added for test timeouts

#Type checking
mypy>=0.910,<2.0.0
types-PyYAML>=6.0.0  # Added for YAML type hints
types-setuptools>=57.0.0  # Added for setuptools type hints

#Development tools
black>=21.6b0,<24.0.0
flake8>=3.9.0,<7.0.0
isort>=5.9.0,<6.0.0
pre-commit>=2.15.0  # Added for git hooks

#Additional required dependencies
tqdm>=4.62.0,<5.0.0  # Added for progress bars
joblib>=1.0.0,<2.0.0  # Added for parallel processing
h5py>=3.1.0,<4.0.0  # Added for data storage
numba>=0.54.0,<0.58.0  # Added for performance optimization